Basic loop ready (without eat countin)

+++ include/philo_stats.h
@@ -3,10 +3,10 @@
 
 typedef struct s_philo_stats
 {
-    long time_to_die;
-    long time_to_eat;
-    long time_to_sleep;
-    long time_to_think;
+    long long time_to_die;
+    long long time_to_eat;
+    long long time_to_sleep;
+    long long time_to_think;
 } t_philo_stats;
 
 #endif
